refactor: modularize extended builtins
