+++ src/ext_builtins/CMakeLists.txt
@@ -1,0 +1,34 @@
+option(ENABLE_EXT_BUILTIN_MATH "Build math extended builtins" ON)
+option(ENABLE_EXT_BUILTIN_STRINGS "Build string extended builtins" ON)
+option(ENABLE_EXT_BUILTIN_SYSTEM "Build system extended builtins" ON)
+option(ENABLE_EXT_BUILTIN_USER "Build user extended builtins" ON)
+
+set(EXT_BUILTIN_SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/register.c)
+set(EXT_BUILTIN_DEFS)
+
+if(ENABLE_EXT_BUILTIN_MATH)
+    list(APPEND EXT_BUILTIN_DEFS ENABLE_EXT_BUILTIN_MATH)
+    add_subdirectory(math)
+    list(APPEND EXT_BUILTIN_SOURCES ${MATH_BUILTIN_SOURCES})
+endif()
+
+if(ENABLE_EXT_BUILTIN_STRINGS)
+    list(APPEND EXT_BUILTIN_DEFS ENABLE_EXT_BUILTIN_STRINGS)
+    add_subdirectory(strings)
+    list(APPEND EXT_BUILTIN_SOURCES ${STRING_BUILTIN_SOURCES})
+endif()
+
+if(ENABLE_EXT_BUILTIN_SYSTEM)
+    list(APPEND EXT_BUILTIN_DEFS ENABLE_EXT_BUILTIN_SYSTEM)
+    add_subdirectory(system)
+    list(APPEND EXT_BUILTIN_SOURCES ${SYSTEM_BUILTIN_SOURCES})
+endif()
+
+if(ENABLE_EXT_BUILTIN_USER)
+    list(APPEND EXT_BUILTIN_DEFS ENABLE_EXT_BUILTIN_USER)
+    file(GLOB USER_BUILTIN_SOURCES CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/user/*.c)
+    list(APPEND EXT_BUILTIN_SOURCES ${USER_BUILTIN_SOURCES})
+endif()
+
+set(EXT_BUILTIN_SOURCES ${EXT_BUILTIN_SOURCES} PARENT_SCOPE)
+set(EXT_BUILTIN_DEFS ${EXT_BUILTIN_DEFS} PARENT_SCOPE)
