+++ src/ext_builtins/strings/register.c
@@ -1,0 +1,5 @@
+#include "backend_ast/builtin.h"
+
+void registerReverseStringBuiltin(void);
+
+void pascal_ext_strings_init(void) { registerReverseStringBuiltin(); }
