+++ src/ext_builtins/system/CMakeLists.txt
@@ -1,0 +1,6 @@
+set(SYSTEM_BUILTIN_SOURCES
+    ${CMAKE_CURRENT_SOURCE_DIR}/fileexists.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/getpid.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/swap.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/register.c
+    PARENT_SCOPE)
