+++ src/ext_builtins/system/getpid.c
@@ -1,0 +1,14 @@
+#include "backend_ast/builtin.h"
+#include "core/utils.h"
+#include <unistd.h>
+
+static Value vmBuiltinGetPid(struct VM_s *vm, int arg_count, Value *args) {
+  (void)vm;
+  (void)args;
+  return arg_count == 0 ? makeInt(getpid()) : makeInt(-1);
+}
+
+void registerGetPidBuiltin(void) {
+  registerBuiltinFunction("GetPid", AST_FUNCTION_DECL, NULL);
+  registerVmBuiltin("getpid", vmBuiltinGetPid);
+}
