+++ src/ext_builtins/user/register.c
@@ -1,0 +1,1 @@
+void pascal_ext_user_init(void) {}
